:truck: move CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING to prescription campaign constants
