+++ api/src/prescription/campaign/infrastructure/repositories/campaign-profiles-collection-participation-summary-repository.js
@@ -1,13 +1,13 @@
 import chunk from 'lodash/chunk.js';
 import isBoolean from 'lodash/isBoolean.js';
 
-import { CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING } from '../../../../shared/constants.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import * as placementProfileService from '../../../../shared/domain/services/placement-profile-service.js';
 import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import { filterByFullName } from '../../../../shared/infrastructure/utils/filter-utils.js';
 import { fetchPage } from '../../../../shared/infrastructure/utils/knex-utils.js';
 import { PromiseUtils } from '../../../../shared/infrastructure/utils/promise-utils.js';
+import { CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING } from '../../domain/constants.js';
 import { CampaignProfilesCollectionParticipationSummary } from '../../domain/read-models/CampaignProfilesCollectionParticipationSummary.js';
 
 async function findPaginatedByCampaignId(campaignId, page = {}, filters = {}) {
