+++ api/src/prescription/campaign/infrastructure/serializers/csv/campaign-profiles-collection-export.js
@@ -1,12 +1,10 @@
 import _ from 'lodash';
 
-import {
-  CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING,
-  CONCURRENCY_HEAVY_OPERATIONS,
-} from '../../../../../shared/constants.js';
+import { CONCURRENCY_HEAVY_OPERATIONS } from '../../../../../shared/constants.js';
 import { serializeLine } from '../../../../../shared/infrastructure/helpers/csv.js';
 import { getI18n } from '../../../../../shared/infrastructure/i18n/i18n.js';
 import { PromiseUtils } from '../../../../../shared/infrastructure/utils/promise-utils.js';
+import { CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING } from '../../../domain/constants.js';
 import { CampaignProfilesCollectionResultLine } from '../../exports/campaigns/campaign-profiles-collection-result-line.js';
 
 class CampaignProfilesCollectionExport {
